refactor the execute command task

+++ src/task/gtest/db/db_put_task-utest.cpp
@@ -2,11 +2,12 @@
 //
 
 #include "task/db/db_put_task.h"
+#include "doim/db/db_key.hpp"
+#include "doim/db/db_value.hpp"
 #include "doim/manager.h"
 #include "err/err.h"
 #include "err/gtest/err.h"
-#include <gtest/gtest-message.h>
-#include <gtest/gtest.h>
+#include "gtest/framework.h"
 #include <memory>
 
 class DbPutTaskTest : public ::testing::Test
@@ -30,7 +31,7 @@ TEST_F(DbPutTaskTest, simple)
     auto key = std::make_shared<doim::DbKey>("key");
     key = doim::gManager->unique(key);
 
-    uint64_t crc = 5;
+    int crc = 5;
     auto value = std::make_shared<doim::DbValue>(crc);
     auto task = std::make_shared<task::DbPutTask>(key, value);
